Implement the main functionality in the parent contract
