+++ packages/protocol/contracts/interfaces/IConvenienceUtils.sol
@@ -1,0 +1,38 @@
+// SPDX-License-Identifier: MIT
+pragma solidity 0.8.6;
+
+interface IConvenienceUtils {
+    event SetAirnodeAnnouncement(
+        address airnode,
+        string xpub,
+        address[] authorizers
+    );
+
+    function setAirnodeAnnouncement(
+        string calldata xpub,
+        address[] calldata authorizers
+    ) external;
+
+    function getAirnodeAnnouncement(address airnode)
+        external
+        view
+        returns (string memory xpub, address[] memory authorizers);
+
+    function checkAuthorizationStatus(
+        address[] calldata authorizers,
+        address airnode,
+        bytes32 requestId,
+        bytes32 endpointId,
+        address sponsor,
+        address requester
+    ) external view returns (bool status);
+
+    function checkAuthorizationStatuses(
+        address[] calldata authorizers,
+        address airnode,
+        bytes32[] calldata requestIds,
+        bytes32[] calldata endpointIds,
+        address[] calldata sponsors,
+        address[] calldata requesters
+    ) external view returns (bool[] memory statuses);
+}
