--- packages/protocol/contracts/interfaces/IRequestUtils.sol
@@ -1,102 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity 0.8.6;
-
-import "./ITemplateUtils.sol";
-
-interface IRequestUtils is ITemplateUtils {
-    event SetSponsorshipStatus(
-        address indexed sponsor,
-        address indexed requester,
-        bool sponsorshipStatus
-    );
-
-    event MadeTemplateRequest(
-        address indexed airnode,
-        bytes32 indexed requestId,
-        uint256 requesterRequestCount,
-        uint256 chainId,
-        address requester,
-        bytes32 templateId,
-        address sponsor,
-        address sponsorWallet,
-        address fulfillAddress,
-        bytes4 fulfillFunctionId,
-        bytes parameters
-    );
-
-    event MadeFullRequest(
-        address indexed airnode,
-        bytes32 indexed requestId,
-        uint256 requesterRequestCount,
-        uint256 chainId,
-        address requester,
-        bytes32 endpointId,
-        address sponsor,
-        address sponsorWallet,
-        address fulfillAddress,
-        bytes4 fulfillFunctionId,
-        bytes parameters
-    );
-
-    event FulfilledRequest(
-        address indexed airnode,
-        bytes32 indexed requestId,
-        uint256 statusCode,
-        bytes data
-    );
-
-    event FailedRequest(address indexed airnode, bytes32 indexed requestId);
-
-    function setSponsorshipStatus(address requester, bool sponsorshipStatus)
-        external;
-
-    function makeTemplateRequest(
-        bytes32 templateId,
-        address sponsor,
-        address sponsorWallet,
-        address fulfillAddress,
-        bytes4 fulfillFunctionId,
-        bytes calldata parameters
-    ) external returns (bytes32 requestId);
-
-    function makeFullRequest(
-        address airnode,
-        bytes32 endpointId,
-        address sponsor,
-        address sponsorWallet,
-        address fulfillAddress,
-        bytes4 fulfillFunctionId,
-        bytes calldata parameters
-    ) external returns (bytes32 requestId);
-
-    function fulfill(
-        bytes32 requestId,
-        address airnode,
-        uint256 statusCode,
-        bytes calldata data,
-        address fulfillAddress,
-        bytes4 fulfillFunctionId
-    ) external returns (bool callSuccess, bytes memory callData);
-
-    function fail(
-        bytes32 requestId,
-        address airnode,
-        address fulfillAddress,
-        bytes4 fulfillFunctionId
-    ) external;
-
-    function sponsorToRequesterToSponsorshipStatus(
-        address sponsor,
-        address requester
-    ) external view returns (bool sponsorshipStatus);
-
-    function requesterToRequestCountPlusOne(address requester)
-        external
-        view
-        returns (uint256 requestCountPlusOne);
-
-    function requestWithIdHasFailed(bytes32 requestId)
-        external
-        view
-        returns (bool hasFailed);
-}
